Create RecentPosts block
The RecentPosts block shows new CategoryPages for a particular category
or CategoryPage type or both.
